Widen live raw block-state identities
Garden fences, gates, and more crop states exceed the original
one-byte raw block catalogue. Widen the compact generated/live identity
lane to u16 without changing canonical u32 palette-packed state IDs.

Encode raw worker frames as little-endian u16 and write structure
placement snapshots as format 7 while retaining format-6 reads. Cover
generation, lighting, terrain views, and above-255 round trips.

Topic: functional-kitchen-gardens
Topic: wheat-farming

+++ native/crates/mclone-scene/src/session.rs
@@ -6900,7 +6900,7 @@ fn reconcile_observer_preview_entry(
         |pos| {
             client
                 .block_state_at_block_pos(pos)
-                .and_then(|state| u8::try_from(state.0).ok())
+                .and_then(|state| u16::try_from(state.0).ok())
         },
         |chunk| client.chunk_snapshot(chunk).is_some(),
     ) else {
